+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed bug when formatting TimeAttribute.
 * Section.direction, Intersection.direction, SideArea.direction changed from 
 String to Direction.
+* RoadSystem.roadCategory, RoadSystem.phase changed from String to enum.
 
 ## 1.2.2
 * Fixed NPE when Roadsystem does not have number.